auth/requirements.txt
aniso8601==9.0.1
attrs==22.1.0
captcha==0.4
click==8.1.3
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
Deprecated==1.2.13
Flask==2.1.2
Flask-Limiter==2.6.2
Flask-Migrate==3.1.0
flask-restx==0.5.1
flask-session-captcha==1.3.0
Flask-Sessionstore==0.4.5
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.9.1
mock==4.0.3
MarkupSafe==2.1.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-jaeger
packaging==21.3
pika==1.3.0
pika-stubs==0.1.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.9.1
pymongo==4.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-mock==3.8.2
PyYAML==6.0
redis==4.3.4
repackage==0.7.3
requests==2.28.1
responses==0.21.0
SQLAlchemy==1.4.40
pytz==2022.2
six==1.16.0
sjwt==0.0.6
types-psycopg2==2.9.21.1
types-requests==2.28.11.2
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.1.2
wrapt==1.14.1
w3lib==2.0.1
zope.event==4.5.0
zope.interface==5.4.0


